Cast payment values to int and update CSS margin
Casts payment-related values to integers in app.py to ensure correct data types for receipt and payment processing. Adds margin-bottom to .success__btn in success.css for improved button spacing.

+++ app.py
@@ -1394,7 +1394,7 @@ def payment_callback():
                 "task": 1,
                 "cashier": "Рецепція центру Довженка",
                 "receipt": {
-                    "sum": sum,
+                    "sum": int(sum),
                     "comment_down": comments,
                     "rows": [
                         {
@@ -1402,16 +1402,16 @@ def payment_callback():
                             "code": "100",
                             "code2": "",
                             "name": "Квиток",
-                            "cnt": sum/price,
-                            "price":price,
+                            "cnt": int(sum/price),
+                            "price":int(price),
                             "taxgrp": 5,
                         },
                     ],
                     "pays": [
                         {
                             "type": 17,
-                            "sum": sum,
-                            "change": 0.00,
+                            "sum": int(sum),
+                            "change": 0,
                             "comment": {comments},
                             "currency": payload.get("currency")
                         }
